[refactor] DB Analyzer
